+++ src/modules/util/excel/excel.module.ts
@@ -5,7 +5,7 @@ import { JwtService } from '@nestjs/jwt';
 import { DatabaseService } from 'src/modules/core/database/mysql/database.service';
 import { ExcelBuilderService } from 'src/modules/util/excel/services/excel.builder.service';
 import { QueueModule } from 'src/modules/core/queue/queue.module';
-import { ZipService } from 'src/services/zip/zip.service';
+import { ZipService } from 'src/services/utils/zip.service';
 import { EmailService } from 'src/services/email/email.service';
 import { EmailConfig } from 'src/settings';
 import { CacheModule } from '@nestjs/cache-manager';
